+++ frontend/app-development/features/processEditor/handlers/OnProcessTaskAddHandler.ts
@@ -1,16 +1,14 @@
 import { PaymentPolicyBuilder } from '../../../utils/policy';
 import type { OnProcessTaskEvent } from '@altinn/process-editor/types/OnProcessTask';
 import type { Policy } from 'app-shared/types/Policy';
-import {
-  getDataTypeIdFromBusinessObject,
-  getReceiptPdfDataTypeIdFromBusinessObject,
-} from '@altinn/process-editor/utils/hookUtils/hookUtils';
 import type {
   AddLayoutSetMutation,
   AddLayoutSetMutationPayload,
 } from '../../../hooks/mutations/useAddLayoutSetMutation';
+import { StudioModeler } from '@altinn/process-editor/utils/bpmnModeler/StudioModeler';
 
 export class OnProcessTaskAddHandler {
+  private readonly studioModeler = new StudioModeler();
   constructor(
     private readonly org: string,
     private readonly app: string,
@@ -58,21 +56,15 @@ export class OnProcessTaskAddHandler {
   private handlePaymentTaskAdd(taskMetadata: OnProcessTaskEvent): void {
     this.addLayoutSet(this.createLayoutSetConfig(taskMetadata.taskEvent));
 
-    const dataTypeId = getDataTypeIdFromBusinessObject(
-      taskMetadata.taskType,
-      taskMetadata.taskEvent.element.businessObject,
-    );
-
+    const dataTypeId = this.studioModeler.getDataTypeIdFromBusinessObject(taskMetadata.taskType);
     this.addDataTypeToAppMetadata({
       dataTypeId,
       taskId: taskMetadata.taskEvent.element.id,
     });
 
-    const receiptPdfDataTypeId = getReceiptPdfDataTypeIdFromBusinessObject(
+    const receiptPdfDataTypeId = this.studioModeler.getReceiptPdfDataTypeIdFromBusinessObject(
       taskMetadata.taskType,
-      taskMetadata.taskEvent.element.businessObject,
     );
-
     this.addDataTypeToAppMetadata({
       dataTypeId: receiptPdfDataTypeId,
       taskId: taskMetadata.taskEvent.element.id,
@@ -96,11 +88,7 @@ export class OnProcessTaskAddHandler {
    * @private
    */
   private handleSigningTaskAdd(taskMetadata: OnProcessTaskEvent): void {
-    const dataTypeId = getDataTypeIdFromBusinessObject(
-      taskMetadata.taskType,
-      taskMetadata.taskEvent.element.businessObject,
-    );
-
+    const dataTypeId = this.studioModeler.getDataTypeIdFromBusinessObject(taskMetadata.taskType);
     this.addDataTypeToAppMetadata({
       dataTypeId,
       taskId: taskMetadata.taskEvent.element.id,
