+++ frontend/app-development/features/processEditor/handlers/OnProcessTaskRemoveHandler.ts
@@ -1,14 +1,12 @@
 import type { Policy } from 'app-shared/types/Policy';
 import type { OnProcessTaskEvent } from '@altinn/process-editor/types/OnProcessTask';
 import { PaymentPolicyBuilder } from '../../../utils/policy';
-import {
-  getDataTypeIdFromBusinessObject,
-  getLayoutSetIdFromTaskId,
-  getReceiptPdfDataTypeIdFromBusinessObject,
-} from '@altinn/process-editor/utils/hookUtils/hookUtils';
 import type { LayoutSets } from 'app-shared/types/api/LayoutSetsResponse';
+import { StudioModeler } from '@altinn/process-editor/utils/bpmnModeler/StudioModeler';
+import { getLayoutSetIdFromTaskId } from '../bpmnHandlerUtils/bpmnHandlerUtils';
 
 export class OnProcessTaskRemoveHandler {
+  private readonly studioModeler = new StudioModeler();
   constructor(
     private readonly org: string,
     private readonly app: string,
@@ -60,20 +58,14 @@ export class OnProcessTaskRemoveHandler {
    * @private
    */
   private handlePaymentTaskRemove(taskMetadata: OnProcessTaskEvent): void {
-    const dataTypeId = getDataTypeIdFromBusinessObject(
-      taskMetadata.taskType,
-      taskMetadata.taskEvent.element.businessObject,
-    );
-
+    const dataTypeId = this.studioModeler.getDataTypeIdFromBusinessObject(taskMetadata.taskType);
     this.deleteDataTypeFromAppMetadata({
       dataTypeId,
     });
 
-    const receiptPdfDataTypeId = getReceiptPdfDataTypeIdFromBusinessObject(
+    const receiptPdfDataTypeId = this.studioModeler.getReceiptPdfDataTypeIdFromBusinessObject(
       taskMetadata.taskType,
-      taskMetadata.taskEvent.element.businessObject,
     );
-
     this.deleteDataTypeFromAppMetadata({
       dataTypeId: receiptPdfDataTypeId,
     });
@@ -109,11 +101,7 @@ export class OnProcessTaskRemoveHandler {
    * @private
    */
   private handleSigningTaskRemove(taskMetadata: OnProcessTaskEvent): void {
-    const dataTypeId = getDataTypeIdFromBusinessObject(
-      taskMetadata.taskType,
-      taskMetadata.taskEvent.element.businessObject,
-    );
-
+    const dataTypeId = this.studioModeler.getDataTypeIdFromBusinessObject(taskMetadata.taskType);
     this.deleteDataTypeFromAppMetadata({
       dataTypeId,
     });
