+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigContent/EditActions/ActionsEditor/ActionsEditor.test.tsx
@@ -5,10 +5,10 @@ import { render, screen, waitFor } from '@testing-library/react';
 import { BpmnContext } from '../../../../../contexts/BpmnContext';
 import { ActionsEditor, type ActionsEditorProps } from './ActionsEditor';
 import { mockBpmnContextValue } from '../../../../../../test/mocks/bpmnContextMock';
-import { BpmnActionModeler, type Action } from '../../../../../utils/bpmn/BpmnActionModeler';
+import { BpmnActionModeler, type Action } from '../../../../../utils/bpmnModeler/BpmnActionModeler';
 import { BpmnConfigPanelFormContextProvider } from '../../../../../contexts/BpmnConfigPanelContext';
 
-jest.mock('../../../../../utils/bpmn/BpmnActionModeler');
+jest.mock('../../../../../utils/bpmnModeler/BpmnActionModeler');
 
 const actionElementMock: Action = {
   $type: 'altinn:Action',
