+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigContent/EditActions/ActionsEditor/ActionsEditor.tsx
@@ -13,7 +13,7 @@ import { CheckmarkIcon } from '@studio/icons';
 import { CustomActions } from './CustomActions';
 import { PredefinedActions } from './PredefinedActions';
 import { useBpmnContext } from '../../../../../contexts/BpmnContext';
-import { type Action, BpmnActionModeler } from '../../../../../utils/bpmn/BpmnActionModeler';
+import { type Action, BpmnActionModeler } from '../../../../../utils/bpmnModeler/BpmnActionModeler';
 import { getPredefinedActions, isActionRequiredForTask } from '../../../../../utils/processActions';
 import classes from './ActionsEditor.module.css';
