+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigContent/EditActions/ActionsEditor/CustomActions/CustomActions.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { StudioTextfield } from '@studio/components';
 import { HelpText, Switch } from '@digdir/design-system-react';
 import { useDebounce } from 'app-shared/hooks/useDebounce';
-import { BpmnActionModeler, ActionType } from '../../../../../../utils/bpmn/BpmnActionModeler';
-import type { Action } from '../../../../../../utils/bpmn/BpmnActionModeler';
+import { BpmnActionModeler, ActionType } from '../../../../../../utils/bpmnModeler/BpmnActionModeler';
+import type { Action } from '../../../../../../utils/bpmnModeler/BpmnActionModeler';
 import { useActionHandler } from '../hooks/useOnActionChange';
 import { getPredefinedActions } from '../../../../../../utils/processActions';
 import { useBpmnContext } from '../../../../../../contexts/BpmnContext';
