+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigContent/EditActions/ActionsEditor/PredefinedActions/PredefinedActions.tsx
@@ -4,7 +4,7 @@ import { StudioNativeSelect } from '@studio/components';
 import { useActionHandler } from '../hooks/useOnActionChange';
 import { useBpmnContext } from '../../../../../../contexts/BpmnContext';
 import { getPredefinedActions, isActionAvailable } from '../../../../../../utils/processActions';
-import { type Action, BpmnActionModeler } from '../../../../../../utils/bpmn/BpmnActionModeler';
+import { type Action, BpmnActionModeler } from '../../../../../../utils/bpmnModeler/BpmnActionModeler';
 
 type PredefinedActionsProps = {
   actionElement: Action;
