+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigContent/EditActions/ActionsEditor/hooks/useOnActionChange.test.tsx
@@ -3,10 +3,10 @@ import { renderHook } from '@testing-library/react';
 import { useActionHandler } from './useOnActionChange';
 import { BpmnContext } from '../../../../../../contexts/BpmnContext';
 import { mockBpmnContextValue } from '../../../../../../../test/mocks/bpmnContextMock';
-import { type Action, BpmnActionModeler } from '../../../../../../utils/bpmn/BpmnActionModeler';
+import { type Action, BpmnActionModeler } from '../../../../../../utils/bpmnModeler/BpmnActionModeler';
 import { BpmnConfigPanelFormContextProvider } from '../../../../../../contexts/BpmnConfigPanelContext';
 
-jest.mock('../../../../../../utils/bpmn/BpmnActionModeler');
+jest.mock('../../../../../../utils/bpmnModeler/BpmnActionModeler');
 
 const actionElementMock: Action = {
   $type: 'altinn:Action',
