+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigSequenceFlow/ConfigSequenceFlow.tsx
@@ -9,7 +9,7 @@ import {
 import { PlusIcon } from '@studio/icons';
 import { useBpmnContext } from '../../../contexts/BpmnContext';
 import { Paragraph } from '@digdir/design-system-react';
-import { BpmnExpressionModeler } from '../../../utils/bpmn/BpmnExpressionModeler';
+import { BpmnExpressionModeler } from '../../../utils/bpmnModeler/BpmnExpressionModeler';
 import { useExpressionTexts } from 'app-shared/components/Expression/useExpressionTexts';
 import { useTranslation } from 'react-i18next';
 import classes from './ConfigSequenceFlow.module.css';
