+++ frontend/packages/process-editor/src/hooks/useBpmnViewer.ts
@@ -3,7 +3,7 @@ import { useRef, useEffect, useState } from 'react';
 import BpmnJS from 'bpmn-js/dist/bpmn-navigated-viewer.development.js';
 import { useBpmnContext } from '../contexts/BpmnContext';
 import type { BpmnViewerError } from '../types/BpmnViewerError';
-import { getBpmnViewerDetailsFromBusinessObject } from '../utils/hookUtils';
+import { getBpmnViewerDetailsFromBusinessObject } from '../utils/bpmnObjectBuilders';
 
 // Wrapper around bpmn-js to Reactify it
