+++ frontend/packages/process-editor/src/utils/bpmnObjectBuilders/bpmnObjectBuilders.test.ts
@@ -9,11 +9,9 @@ import type { BpmnTaskType } from '../../types/BpmnTaskType';
 import {
   getBpmnEditorDetailsFromBusinessObject,
   getBpmnViewerDetailsFromBusinessObject,
-  getLayoutSetIdFromTaskId,
-} from './hookUtils';
-import { mockBpmnDetails } from '../../../test/mocks/bpmnDetailsMock';
+} from './bpmnObjectBuilders';
 
-describe('hookUtils', () => {
+describe('bpmnObjectBuilders', () => {
   afterEach(jest.clearAllMocks);
 
   describe('getBpmnViewerDetailsFromBusinessObject', () => {
@@ -84,7 +82,7 @@ describe('hookUtils', () => {
       expect(bpmnDetails.taskType).toEqual(mockTaskTypeData);
     });
 
-    it('returns taskType with value "null" when etensionElements are not present', () => {
+    it('returns taskType with value "null" when extensionElements are not present', () => {
       const bpmnBusinessObject: BpmnBusinessObjectEditor = {
         ...mockBpmnBusinessObject,
         extensionElements: undefined,
@@ -96,7 +94,7 @@ describe('hookUtils', () => {
       expect(bpmnDetails.taskType).toBeNull();
     });
 
-    it('returns taskType with value "null" when etensionElements.values are not present', () => {
+    it('returns taskType with value "null" when extensionElements.values are not present', () => {
       const bpmnBusinessObject: BpmnBusinessObjectEditor = {
         ...mockBpmnBusinessObject,
         extensionElements: {
@@ -109,31 +107,6 @@ describe('hookUtils', () => {
       expect(bpmnDetails.name).toEqual(mockName);
       expect(bpmnDetails.type).toEqual(mockTypeTask);
       expect(bpmnDetails.taskType).toBeNull();
-    });
-  });
-
-  const layoutSets = {
-    sets: [
-      { id: 'layoutSet1', tasks: ['task1'] },
-      { id: 'layoutSet2', tasks: ['task2'] },
-    ],
-  };
-
-  describe('getLayoutSetIdFromTaskId', () => {
-    it('should return the layout set id corresponding to the task id', () => {
-      const result = getLayoutSetIdFromTaskId('task1', layoutSets);
-      expect(result).toBe('layoutSet1');
-    });
-
-    it('should return undefined if task id does not exist in any layout set', () => {
-      const result = getLayoutSetIdFromTaskId(mockBpmnDetails.id, layoutSets);
-      expect(result).toBeUndefined();
-    });
-
-    it('should return undefined if layout sets are empty', () => {
-      const layoutSets = { sets: [] };
-      const result = getLayoutSetIdFromTaskId(mockBpmnDetails.id, layoutSets);
-      expect(result).toBeUndefined();
     });
   });
 });
